Меняет html-структуру заголовков внтури статьи
Fix: #622
