+++ src/transforms/code-transform.js
@@ -94,6 +94,7 @@ module.exports = function(window) {
       'index-group-list__link': 'index-group-list__code',
       'header__title': 'header__title-code',
       'article__description': 'article__description-code',
+      'article-heading': 'article-heading__code'
     }
 
     for (const [parentClass, codeClass] of Object.entries(classMap)) {
